requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.1
pandas>=2.0.0
python-docx>=0.8.11
requests>=2.28.0
